+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Fixed func LinkAll and ILinkAll to exclude words in existing A elements
+
 ## [0.16.0] 2021-03-04
 
 - Add CSS.SaveAs and related methods for saving to file