+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.17.0] 2021-04-09
 
 - Add package sweb for sequential writing style
 - Removed CheckLinks channel argument